refactor history code to history.h
bench 1281912
